requirements.txt
APScheduler==3.6.3
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
parso==0.8.1
pycodestyle==2.7.0
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
